Disable ELF metadata because it creeps into everything (LP: #2109333)
This is only a temporary workaround until dpkg is fixed to set it
through CFLAGS/LDFLAGS, at which point someone will have to redo this
properly so that ld.so and libc.so.6 both have the metadata but not
crt1.o and friends. This particular workaround results in NONE of those
having the data.